+++ src/app/styles/reset.css
@@ -1,4 +1,3 @@
-
 /* prettier-ignore */
 html, body, div, span, applet, object, iframe,
 h1, h2, h3, h4, h5, h6, p, blockquote, pre,
@@ -47,17 +46,16 @@ q::after {
 table {
   border-collapse: collapse;
   border-spacing: 0;
+}
 
 * {
   margin: 0;
   padding: 0;
   box-sizing: border-box;
-
 }
 
 button {
   cursor: pointer;
-
 }
 
 a {
