+++ src/shared/ui/form-field/FormField.tsx
@@ -30,7 +30,7 @@ export function FormField({
         errors,
       }}
     >
-      {children}
+      <div className="flex flex-col gap-2.5">{children}</div>
     </FormFieldContext.Provider>
   )
 }
@@ -38,22 +38,33 @@ export function FormField({
 export function FormLabel({ children }: LabelProps) {
   const { name } = useFormField()
   return (
-    <label htmlFor={name} className="text-base">
+    <label htmlFor={name} className="text-[16px] text-white">
       {children}
     </label>
   )
 }
 export const Input = forwardRef<HTMLInputElement, InputProps>((props, ref) => {
   const { name, register, errors } = useFormField()
-  const errorClassName = errors ? 'border-red' : ''
-  return <input id={name} autoComplete={name} {...register(name)} {...props} />
+  const error = errors[name]
+  const errorClassName = error ? 'border-red' : 'border-black-70'
+  return (
+    <input
+      id={name}
+      autoComplete={name}
+      {...register(name)}
+      {...props}
+      className={`h-[70px] text-gray-50 rounded-lg px-5 placeholder:text-black-30 bg-black-60 border caret-white ${errorClassName}`}
+    />
+  )
 })
 
 export function FormErrorMessage() {
   const { name, errors } = useFormField()
   const error = errors[name]
 
-  if (!error) return null
-
-  return <p>{error.message as React.ReactNode}</p>
+  return (
+    <p className="h-1 text-red text-sm">
+      {error ? (error.message as React.ReactNode) : null}
+    </p>
+  )
 }
