+++ src/shared/ui/form/Form.tsx
@@ -1,0 +1,13 @@
+interface FormProps {
+  onSubmit: (event: React.FormEvent<HTMLFormElement>) => void
+  children: React.ReactNode
+  className?: string
+}
+
+export function Form({ onSubmit, children, className }: FormProps) {
+  return (
+    <form onSubmit={onSubmit} className={className}>
+      {children}
+    </form>
+  )
+}
